# api/swagger-spec/swagger.json
{
  "schemes": [
    "http"
  ],
  "openapi": "3.0.3",
  "info": {
    "description": "The purpose of this application is to provide access control mechanism",
    "title": "Policy REST API.",
    "contact": {
      "name": "echaoeoen",
      "email": "andyepso@gmail.com"
    },
    "version": "0.0.1"
  },
  "basePath": "/",
  "paths": {
    "/enforce/tenant/{tenant}": {
      "post": {
        "description": "Provide answer whether access is permitted or not",
        "tags": [
          "enforcer"
        ],
        "operationId": "enforce",
        "parameters": [
          {
            "description": "Tenant domain",
            "name": "tenant",
            "schema": {
              "type": "string"
            },
            "in": "path",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/EnforceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "access is permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Policy"
                }
              }
            }
          },
          "401": {
            "description": "access is forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Policy"
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "EnforceRequest": {
      "description": "EnforceRequest represent http request to enforce API",
      "type": "object",
      "required": [
        "subject",
        "action",
        "resource"
      ],
      "properties": {
        "action": {
          "description": "the action taken by subject",
          "type": "string",
          "format": "string"
        },
        "resource": {
          "description": "the resource taking the action from subject",
          "type": "string",
          "format": "string"
        },
        "subject": {
          "description": "subject who will take the action",
          "type": "string",
          "format": "string"
        }
      }
    },
...
    "Policy": {
      "description": "Policy represent access control in the form of \"a Subject is [allowed|denied] to do an Action on an Object\" with additional metadata",
      "type": "object",
      "properties": {
        "action": {
          "type": "string",
          "format": "string"
        },
        "active": {
          "type": "boolean"
        },
        "effect": {
          "type": "string",
          "description": "string of \"allow\" or \"deny\""
        },
        "created": {
          "type": "string",
          "format": "date-time"
        },
        "expire": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "string"
        },
        "resources": {
          "description": "selector",
          "type": "string",
          "format": "string"
        },
        "subject": {
          "type": "string",
          "format": "string"
        },
        "updated": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
  }
}